MEOW-57 backgeocoding optimizations
Use int16_t for raster data.

sincos instead of seperate sin&cos in kernels.

Compute extended amount no longer reallocates metadata for each call and uses atomics for index calculations. This change also fixes a GPU memory leak in the previous version.

Threading refactoring. Coregistration output now supports 4 datasets with 1 band or 1 dataset with 4 bands.

Minor kernel launch parameters changes due optimization or launch failures on CC7.5.

+++ algs/backgeocoding/include/deramp_demod_computation.h
@@ -22,8 +22,8 @@ namespace alus {
 namespace backgeocoding {
 
 cudaError_t LaunchDerampDemod(Rectangle rectangle,
-                              double *slave_i,
-                              double *slave_q,
+                              int16_t *slave_i,
+                              int16_t *slave_q,
                               double *demod_phase,
                               double *demod_i,
                               double *demod_q,
